warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@776 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/MachineDebuggerImpl.cpp
@@ -607,7 +607,7 @@ void MachineDebugger::flushQueuedSettings()
         COMSETTER(LogEnabled)(mLogEnabledQueued);
         mLogEnabledQueued = ~0;
     }
-    if (mVirtualTimeRateQueued != ~0)
+    if (mVirtualTimeRateQueued != ~(uint32_t)0)
     {
         COMSETTER(VirtualTimeRate)(mVirtualTimeRateQueued);
         mVirtualTimeRateQueued = ~0;
